devices to things
