+++ i2ctools/integrationplugini2ctools.json
@@ -7,7 +7,7 @@
             "displayName": "nymea",
             "name": "nymea",
             "id": "2062d64d-3232-433c-88bc-0d33c0ba2ba6",
-            "deviceClasses": [
+            "thingClasses": [
                 {
                     "id": "468dbda3-4364-4255-8060-25e80ab9fc29",
                     "name": "i2cInterface",
