+++ i2ctools/translations/284d7f44-09aa-444e-afc3-5acee8f2ab43-de_DE.ts
@@ -1,13 +1,13 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!DOCTYPE TS>
-<TS version="2.1">
+<TS version="2.1" language="de_DE">
 <context>
     <name>I2cTools</name>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="39"/>
         <source>Address</source>
         <extracomment>The name of the ParamType (ThingClass: i2cReadRegister, Type: thing, ID: {fc0a40ef-6ae1-4adf-9f41-da6dbb37d8b5})</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>Adresse</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="42"/>
@@ -16,13 +16,13 @@
         <extracomment>The name of the ParamType (ThingClass: i2cInterface, EventType: availableDevices, ID: {6a1a5da4-a8ff-4173-8238-6196aecef545})
 ----------
 The name of the StateType ({6a1a5da4-a8ff-4173-8238-6196aecef545}) of ThingClass i2cInterface</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>Vorhandene Geräte</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="48"/>
         <source>Available devices changed</source>
         <extracomment>The name of the EventType ({6a1a5da4-a8ff-4173-8238-6196aecef545}) of ThingClass i2cInterface</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>Vorhandene Geräte geändert</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="51"/>
@@ -37,7 +37,7 @@ The name of the StateType ({726befc0-37cd-47f1-b983-1b5ee87a5bf7}) of ThingClass
 The name of the ParamType (ThingClass: i2cInterface, EventType: connected, ID: {a108bdaf-2067-4ec9-bd45-89c5eecb8038})
 ----------
 The name of the StateType ({a108bdaf-2067-4ec9-bd45-89c5eecb8038}) of ThingClass i2cInterface</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>Verbunden</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="63"/>
@@ -46,13 +46,13 @@ The name of the StateType ({a108bdaf-2067-4ec9-bd45-89c5eecb8038}) of ThingClass
         <extracomment>The name of the EventType ({726befc0-37cd-47f1-b983-1b5ee87a5bf7}) of ThingClass i2cReadRegister
 ----------
 The name of the EventType ({a108bdaf-2067-4ec9-bd45-89c5eecb8038}) of ThingClass i2cInterface</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>Verbunden geändert</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="69"/>
         <source>I2C Tools</source>
         <extracomment>The name of the plugin I2cTools ({284d7f44-09aa-444e-afc3-5acee8f2ab43})</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>I2C Werkzeuge</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="72"/>
@@ -61,19 +61,19 @@ The name of the EventType ({a108bdaf-2067-4ec9-bd45-89c5eecb8038}) of ThingClass
         <extracomment>The name of the ParamType (ThingClass: i2cInterface, Type: thing, ID: {406b9787-62c1-4697-9c97-fe4105f365d8})
 ----------
 The name of the ThingClass ({468dbda3-4364-4255-8060-25e80ab9fc29})</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>I2C Schnittstelle</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="78"/>
         <source>I2C read register</source>
         <extracomment>The name of the ThingClass ({f6c2ab46-3a70-4dbd-aae4-62fb7e984514})</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>I2C Leseregister</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="81"/>
         <source>Register</source>
         <extracomment>The name of the ParamType (ThingClass: i2cReadRegister, Type: thing, ID: {27a86394-d89f-4140-966d-ed8d5eb65f64})</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>Register</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="84"/>
@@ -82,19 +82,19 @@ The name of the ThingClass ({468dbda3-4364-4255-8060-25e80ab9fc29})</extracommen
         <extracomment>The name of the ParamType (ThingClass: i2cReadRegister, EventType: value, ID: {72bf6436-e55e-45c7-b1c7-f459ae587b00})
 ----------
 The name of the StateType ({72bf6436-e55e-45c7-b1c7-f459ae587b00}) of ThingClass i2cReadRegister</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>Wert</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="90"/>
         <source>Value changed</source>
         <extracomment>The name of the EventType ({72bf6436-e55e-45c7-b1c7-f459ae587b00}) of ThingClass i2cReadRegister</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>Wert geändert</translation>
     </message>
     <message>
         <location filename="../../../build-nymea-plugins-Desktop-Debug/i2ctools/plugininfo.h" line="93"/>
         <source>nymea</source>
         <extracomment>The name of the vendor ({2062d64d-3232-433c-88bc-0d33c0ba2ba6})</extracomment>
-        <translation type="unfinished"></translation>
+        <translation>nymea</translation>
     </message>
 </context>
 </TS>
